핸드폰 인증 체크 개발 및 소스 정리

+++ msg-pick-partner/src/main/java/com/msgpick/module/auth/service/AuthService.java
@@ -8,6 +8,7 @@ import com.msgpick.module.partners.domain.Partner;
 import com.msgpick.module.partners.dto.PartnerRegisterRequest;
 import com.msgpick.module.partners.repository.PartnerRepository;
 import com.msgpick.msgpick.global.common.exception.BaseException;
+import com.msgpick.msgpick.global.common.exception.EntityNotFoundException;
 import com.msgpick.msgpick.global.common.response.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -15,8 +16,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 
-@Service
 @RequiredArgsConstructor
+@Service
 public class AuthService {
 
     private final PasswordEncoder passwordEncoder;
@@ -57,8 +58,14 @@ public class AuthService {
     @Transactional
     public void findVerifications(AuthCheckVerifiedRequest request) {
         Auth checkVerifiedNumber = authRepository.findByPhone(request.getPhone())
-                .orElseThrow(() -> new BaseException(ErrorCode.VERIFICATION_NUMBER_NOT_MATCHED));
+                .orElseThrow(() -> new EntityNotFoundException("해당 핸드폰번호가 없습니다 : " + request.getPhone()));
 
-        authRepository.delete(checkVerifiedNumber);
+        if (request.getVerificationCode() == checkVerifiedNumber.getVerificationCode()) {
+            authRepository.delete(checkVerifiedNumber);
+        } else {
+            throw new BaseException(ErrorCode.VERIFICATION_NUMBER_NOT_MATCHED);
+        }
     }
+
 }
+
